Rework the way the tags are assigned to the role
Now that we are using default tags in our providers, we need to rework
the way tags are assigned to the role resource.  This is because the
provider's default tags will be assigned to the role resource, and if
those tags are identical to the tags assigned explicitly then
Terraform returns an error.

The solution is to remove the role_tags variable and replace it with
an additional_role_tags variable.  This variable contains _only_ the
tags to be assigned to the role resource _in addition to_ the
provider's default tags.